chore: move decorators into main package
