+++ catalyst/src/index.ts
@@ -1,5 +1,7 @@
 import {bind} from './bind'
 import {register} from './register'
 import {findTarget, findTargets} from './findtarget'
+import {target, targets} from './target'
+import {controller} from './controller'
 
-export {bind, register, findTarget, findTargets}
+export {bind, register, findTarget, findTargets, target, targets, controller}
