--- decorators/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@catalyst/decorators",
-  "version": "1.0.0",
-  "private": true,
-  "description": "Helpers for creating HTML Elements as Controllers",
-  "license": "MIT",
-  "author": "GitHub Inc.",
-  "contributors": [
-    "Keith Cirkel (https://keithcirkel.co.uk/)",
-    "Kristján Oddsson <koddsson@gmail.com>"
-  ],
-  "type": "module",
-  "main": "lib/index.js",
-  "module": "lib/index.js",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "dependencies": {
-    "@github/catalyst": "^1.0.0"
-  }
-}
