--- decorators/src/controller.ts
@@ -1,11 +1,0 @@
-import {register} from '@github/catalyst'
-import {bindEvents} from './bind'
-
-interface CustomElement {
-  new(): HTMLElement
-}
-
-export function controller(classObject: CustomElement) {
-  bindEvents(classObject)
-  register(classObject)
-}
