--- decorators/src/helpers.ts
@@ -1,2 +1,0 @@
-export {bindEvents} from './bind'
-export {target, targets} from './target'
