--- decorators/src/index.ts
@@ -1,4 +1,0 @@
-import {bindEvents} from './bind'
-import {target, targets} from './target'
-
-export {bindEvents, target, targets}
