--- decorators/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "include": ["src"],
-  "compilerOptions": {
-    "composite": true,
-    "noEmit": false,
-    "rootDir": "src",
-    "outDir": "./lib"
-  },
-  "references": [
-    { "path": "../catalyst" }
-  ]
-}
